configure: Move check for custom libbpf install to a helper
If LIBBPF_DIR environment is set, it should take precedence over
system and git submodule. Move the check to a helper and fix it
to detect an install done via

  cd src
  DESTDIR=/some/path make install

within the libbpf repository.

Such an install may not have the elf and zlib dependencies set,
so refactor those two functions to set the libs in LDLIBS which
is then passed to libbpf_compile_test.

Signed-off-by: David Ahern <dsahern@gmail.com>